+++ CHANGELOG.md
@@ -7,6 +7,9 @@
 - Added files page length to client settings.
 - Added automatic namer for cinematics (videos, audio, subtitles).
 - Added startup check to ensure WTL is always launched from the correct directory.
+- Added support for Blizzard's new version service (TACT Channels), setting is off by default for now.
+- Added button to builds page to clear version service cache (and also reload local build information).
+- Updated version retrieval across WTL to go through TACTSharp's new VersionServices instead.
 - Updated styling of buttons/update check in header.
 - Updated automatic world map naming to more likely pick up on real filenames.
 - Updated TACT key metadata.